Send packet in from action and Answer port desc request

+++ python/horse.pxd
@@ -29,7 +29,7 @@ cdef extern from "sim/sim.h":
     void host_set_intf_ipv4(host *h, uint32_t port_id, uint32_t addr)
     uint64_t host_uuid(const host* h)
     void host_add_app(host *h, uint16_t type)
-    void host_start_app(host *h, uint16_t type, uint32_t start_time, void* args)
+    void host_start_app(host *h, uint16_t type, uint64_t start_time, void* args)
     void host_set_intf_ipv4(host *h, uint32_t port_id, uint32_t addr, uint32_t netmask)
 
     # Topology.h
